More fixes

+++ korge/src/commonMain/kotlin/korlibs/event/Events.kt
@@ -186,8 +186,8 @@ data class Touch(
     var kind: Kind = Kind.FINGER,
     var button: MouseButton = MouseButton.LEFT,
 ) : Extra by Extra.Mixin() {
-    val x: Float get() = p.x
-    val y: Float get() = p.y
+    val x: Double get() = p.x
+    val y: Double get() = p.y
 
     enum class Status { ADD, KEEP, REMOVE }
     enum class Kind { FINGER, MOUSE, STYLUS, ERASER, UNKNOWN }
