+++ korge/src/commonMain/kotlin/korlibs/korge/ui/UIMaterialLayer.kt
@@ -17,7 +17,7 @@ inline fun Container.uiMaterialLayer(
 ): UIMaterialLayer = UIMaterialLayer(size).addTo(this).apply(block)
 
 class MaterialLayerHighlights(val view: View) {
-    class Highlight(var pos: Point, var radiusRatio: Float, var alpha: Float)
+    class Highlight(var pos: Point, var radiusRatio: Double, var alpha: Double)
 
     @PublishedApi internal val highlights = fastArrayListOf<Highlight>()
     private val highlightsActive = fastArrayListOf<Highlight>()
@@ -30,7 +30,7 @@ class MaterialLayerHighlights(val view: View) {
 
     fun addHighlight(pos: Point) {
         removeHighlights()
-        val highlight = Highlight(pos, 0f, 1f)
+        val highlight = Highlight(pos, 0.0, 1.0)
         highlights += highlight
         highlightsActive += highlight
         view.simpleAnimator.tween(highlight::radiusRatio[1.0], V2Callback { view.invalidateRender() }, time = 0.5.seconds, easing = Easing.EASE_IN)
@@ -59,7 +59,7 @@ class UIMaterialLayer(
     @ViewProperty
     var borderColor: RGBA = Colors.BLACK; set(value) { field = value; invalidateRender() }
     @ViewProperty
-    var borderSize: Float = 0f; set(value) { field = value; invalidateRender() }
+    var borderSize: Double = 0.0; set(value) { field = value; invalidateRender() }
 
     //var highlightPos = Point(0.5, 0.5); set(value) { field = value; invalidateRender() }
     //var highlightRadius = 0.0; set(value) { field = value; invalidateRender() }
@@ -68,7 +68,7 @@ class UIMaterialLayer(
     @ViewProperty
     var shadowColor: RGBA = Colors.BLACK.withAf(0.3f); set(value) { field = value; invalidateRender() }
     @ViewProperty
-    var shadowRadius: Float = 10f; set(value) { field = value; invalidateRender() }
+    var shadowRadius: Double = 10.0; set(value) { field = value; invalidateRender() }
     @ViewProperty
     var shadowOffset: Point = Point.ZERO; set(value) { field = value; invalidateRender() }
 
@@ -80,10 +80,10 @@ class UIMaterialLayer(
         renderCtx2d(ctx) { ctx2d ->
             //println("context.multiplyColor=${ctx2d.multiplyColor}")
             ctx2d.materialRoundRect(
-                x = 0f,
-                y = 0f,
-                width = width.toFloat(),
-                height = height.toFloat(),
+                x = 0.0,
+                y = 0.0,
+                width = width,
+                height = height,
                 color = bgColor,
                 radius = radius,
                 shadowOffset = shadowOffset,
@@ -98,10 +98,10 @@ class UIMaterialLayer(
             )
             highlights.fastForEach {
                 ctx2d.materialRoundRect(
-                    x = 0f,
-                    y = 0f,
-                    width = width.toFloat(),
-                    height = height.toFloat(),
+                    x = 0.0,
+                    y = 0.0,
+                    width = width,
+                    height = height,
                     color = Colors.TRANSPARENT,
                     radius = radius,
                     highlightPos = it.pos,
