+++ korge/src/commonMain/kotlin/korlibs/korge/view/filter/DropShadowFilter.kt
@@ -1,10 +1,10 @@
 package korlibs.korge.view.filter
 
 import korlibs.graphics.shader.*
+import korlibs.image.color.*
 import korlibs.korge.render.*
 import korlibs.korge.view.*
 import korlibs.korge.view.property.*
-import korlibs.image.color.*
 import korlibs.math.*
 import korlibs.math.geom.*
 
@@ -43,7 +43,7 @@ open class DropshadowFilter(
         texHeight: Int,
         renderColorMul: RGBA,
         blendMode: BlendMode,
-        filterScale: Float,
+        filterScale: Double,
     ) {
         //println(blur.border)
         blur.radius = blurRadius
